// src/frontend/react-admin/src/generated/models/bureau-request.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Tracking Application
 * Backend APIs pour une application de Tacking Visites
 *
 * OpenAPI spec version: 1.0
 * Contact: donfackduval@gmail.com
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
/**
 * 
 * @export
 * @interface BureauRequest
 */
export interface BureauRequest {
    /**
     * 
     * @type {string}
     * @memberof BureauRequest
     */
    batiment?: string;
    /**
     * 
     * @type {string}
     * @memberof BureauRequest
     */
    etage?: string;
    /**
     * 
     * @type {string}
     * @memberof BureauRequest
     */
    porte?: string;
}