// src/frontend/react-admin/src/generated/models/bureau-response.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Tracking Application
 * Backend APIs pour une application de Tacking Visites
 *
 * OpenAPI spec version: 1.0
 * Contact: donfackduval@gmail.com
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
/**
 * 
 * @export
 * @interface BureauResponse
 */
export interface BureauResponse {
    /**
     * 
     * @type {number}
     * @memberof BureauResponse
     */
    id?: number;
    /**
     * 
     * @type {string}
     * @memberof BureauResponse
     */
    batiment?: string;
    /**
     * 
     * @type {string}
     * @memberof BureauResponse
     */
    etage?: string;
    /**
     * 
     * @type {string}
     * @memberof BureauResponse
     */
    porte?: string;
    /**
     * 
     * @type {Date}
     * @memberof BureauResponse
     */
    createdAt?: Date;
    /**
     * 
     * @type {Date}
     * @memberof BureauResponse
     */
    updatedAt?: Date;
}